✨ feat: small refactors and output change
