+++ src/index.ts
@@ -1,18 +1,8 @@
 import { readJSON } from "fs-extra";
 import path from "path";
 import puppeteer from "puppeteer";
-import {
-  likeButtonSelector,
-  loginButtonSelector,
-  notNowNotificationsButtonSelector,
-  notNowRememberButtonSelector,
-  passwordFieldSelector,
-  post1Selector,
-  post2Selector,
-  post3Selector,
-  usernameFieldSelector,
-} from "./constants";
-import { Settings } from "./types";
+import { likePosts, login } from "./steps";
+import { Change, Settings } from "./types";
 
 const main = async () => {
   let browser;
@@ -23,120 +13,22 @@ const main = async () => {
     );
     browser = await puppeteer.launch({ headless: false });
     page = await browser.newPage();
-    const username = settings.username;
-    const password = settings.password;
 
-    if (!username || !password) {
-      throw new Error("No username or password set!");
-    }
+    await login(settings, page);
 
-    await page.goto("https://www.instagram.com/", {
-      waitUntil: "networkidle2",
-    });
-
-    const usernameField = await page.$(usernameFieldSelector);
-    const passwordField = await page.$(passwordFieldSelector);
-
-    if (!usernameField || !passwordField) {
-      throw new Error("Could not find inputs on login page!");
-    }
-
-    await usernameField.type(username);
-    await passwordField.type(password);
-
-    const loginButton = await page.$(loginButtonSelector);
-    if (!loginButton) {
-      throw new Error("Could not find login button on login page!");
-    }
-    await loginButton.click();
-    await page.waitForNavigation();
-
-    const notNowRememberButton = await page.$(notNowRememberButtonSelector);
-    if (!notNowRememberButton) {
-      throw new Error("Could not find not now button after login page!");
-    }
-    await notNowRememberButton.click();
-    await page.waitForNavigation();
-
-    const notNowNotificationsButton = await page.$(
-      notNowNotificationsButtonSelector
-    );
-    if (!notNowNotificationsButton) {
-      throw new Error(
-        "Could not find not now notitfications button on home page!"
-      );
-    }
-    await notNowNotificationsButton.click();
-
-    const newlyLikedPosts: (string | null)[] = [];
+    const changes: Change[] = [];
 
     for (let i = 0; i < settings.accounts.length; i++) {
       const account = settings.accounts[i];
-      const posts: (string | null)[] = [];
 
-      await page.goto(`https://www.instagram.com/${account.username}/`, {
-        waitUntil: "networkidle2",
-      });
-
-      const post1 = await page.$eval(post1Selector, (element) => {
-        return element.getAttribute("href");
-      });
-      const post2 = await page.$eval(post2Selector, (element) => {
-        return element.getAttribute("href");
-      });
-      const post3 = await page.$eval(post3Selector, (element) => {
-        return element.getAttribute("href");
-      });
-
-      posts.push(post1);
-      posts.push(post2);
-      posts.push(post3);
-
-      await page.goto(`https://www.instagram.com${post1}`, {
-        waitUntil: "networkidle2",
-      });
-      const likeButton1 = await page.$(likeButtonSelector);
-      if (likeButton1) {
-        await likeButton1.click();
-        newlyLikedPosts.push(post1);
-      }
-      await page.screenshot({
-        path: `${account.username}-post1.png`,
-        type: "png",
-      });
-
-      await page.goto(`https://www.instagram.com${post2}`, {
-        waitUntil: "networkidle2",
-      });
-      const likeButton2 = await page.$(likeButtonSelector);
-      if (likeButton2) {
-        await likeButton2.click();
-        newlyLikedPosts.push(post2);
-      }
-      await page.screenshot({
-        path: `${account.username}-post2.png`,
-        type: "png",
-      });
-
-      await page.goto(`https://www.instagram.com${post3}`, {
-        waitUntil: "networkidle2",
-      });
-      const likeButton3 = await page.$(likeButtonSelector);
-      if (likeButton3) {
-        await likeButton3.click();
-        newlyLikedPosts.push(post3);
-      }
-      await page.screenshot({
-        path: `${account.username}-post3.png`,
-        type: "png",
-      });
+      await likePosts(page, changes, account);
     }
 
     await browser.close();
-    console.log({ newlyLikedPosts });
+    console.log(JSON.stringify({ changes }, null, 2));
   } catch (error) {
     browser?.close();
-    page?.screenshot({ path: 'error.png', type: 'png' })
+    page?.screenshot({ path: "error.png", type: "png" });
     console.error(error);
     process.exit(1);
   }
