+++ src/types.ts
@@ -2,6 +2,11 @@ export interface Account {
   username: string;
 }
 
+export interface Change {
+  account: Account;
+  post: string;
+}
+
 export interface Settings {
   accounts: Account[];
   password: string;
